+++ src/index.ts
@@ -1,10 +1,8 @@
 import FactoryGirl from './FactoryGirl.js';
 
+export { default as FactoryGirl } from './FactoryGirl.js';
 export { default as DefaultAdapter } from './adapters/DefaultAdapter.js';
 
 const factory = new FactoryGirl();
-factory.FactoryGirl = FactoryGirl;
-
-export { factory };
 
 export default factory;